Cadastro de Eventos - ajuste nos hooks do provider, grid do ant, modal copiar eventos em arquivo separado
